DuckOverflow/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
requests==2.32.5
cryptography==46.0.2
Pillow>=11.0.0b1
daphne==4.2.1
whitenoise==6.11.0